Change sections check to template
